# sns-calendar-app/apps/api/app/schemas/auth.py
from typing import Optional

from pydantic import BaseModel, EmailStr, Field, SecretStr


class SignupRequest(BaseModel):
    email: EmailStr
    password: SecretStr = Field(min_length=8)
    display_name: Optional[str] = None


class LoginRequest(BaseModel):
    email: EmailStr
    password: SecretStr


class RefreshRequest(BaseModel):
    refresh_token: str


class UserResponse(BaseModel):
    id: str
    email: EmailStr
    display_name: Optional[str] = None
    ui_mode: str = "simple"
    help_mode_enabled: bool = True
    default_org_id: Optional[str] = None


class SessionResponse(BaseModel):
    access_token: str
    refresh_token: str
    expires_in: int
    token_type: str = "bearer"
    user: UserResponse


class MessageResponse(BaseModel):
    message: str